HPCC-9491 - Ensure lock checked in multiple cluster case
If a logical file had multiple Cluster's, the locking
mechanism was bypassed. Consequently, the removal of a file
on one of the clusters that was locked by others, would still
succeed, leaving the lock dangline.

Clear up dfuplus/dfu's use of 'nodelete' which used to remove
the logical file and leave the physicals orphaned, but hasn't
worked for some time.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>