HPCC-HOSTNAME Fix to allow localThor setup to work with hostnames
With a local Thor setup (localThor=true), if hostnames were used
in the environmnent as opposed to "." or IP's, Thor failed to
start for a couple of reasons.
a) the daliadmin command lines in the script expected "." or IP's.
Fixed with addition of server=
b) the hostname failed to match the IP's in the thorgroup with
the consequence that no slaves were started by the start_slaves
scripts. Fixed by using an IP from the group.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>