HPCC-13985 Global loops not marking some subgraphs as global.
If a LOOP is in a global graph, it must mark all subgraphs as
global. It was not marking them global if they contained entirely
local activities.
That caused problems, if a subgraph tried to access a result
created by a dependency that wasn't marked as global. In that
case, the slaves went ahead executed the dependency without
waiting for the master and then might try to fetch the global
result.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>