HPCC-25342 Conditionally redirect to stdout and ensure correct format
In some context (e.g. eclccserver pipe) stderr cannot be used
for logging, as it's reserved and read back from the pipe as errors.
Add a --logToStdOut option that isn't used and will be captured by
standard containerized logging mechanisms.
NB: it can't be used in all contexts, because stdout is used to
capture the result in some context, e.g. when ecl run creates
an archive.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>