HPCC-27640 Ensure global sort stats are merge correctly.
Global sort was merging it's sorter spill stats per call to
serializeStats, resulting in inflated aggregate figures.

Also, introduce a helper and members to standardize the approach
used for previous gathered stats. and the fetching and merging
of new active stats.
Use these helpers to simplify similar process in diskwrite

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>