too much
Difficult to restructure.
Trying to disassociated the epoll notifySelected handling, from
a thread-like-pool which perform the reading, but it is quite
interwined with the channels (which originall add themselves (well
a CMPPacketReader owned handler) to the notifySelected callback.

In a thread notifyHandler scenario, the initial setup (after accept)
will be handled on a thread, as well as the future epoll notifies
as new data arrives.
This needs to be processed then handed over to the channel.

However, the channels initiate the attachSocket (channel 1st live),
as well as the closeSocket (channel no longer live/listening).

The single notifySelected/threads, need to know whether the
notifiction of new data is for a new socket (after accept), or
for a 'live' channel.